requirements.txt
# PostForge — Instagram Automation Engine
# Core
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# AI — Copy Generation
anthropic==0.40.0

# Image Processing
Pillow==11.1.0

# File Watcher
watchdog==6.0.0

# Templates
jinja2==3.1.5

# HTTP Client
httpx==0.28.1

# Utilities
python-slugify==8.0.4
arrow==1.3.0
loguru==0.7.3

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
